Add civic upgrades, wire difficulty to AI income, merge splash into start screen
Civic upgrades (spend a structure's populace):
- One-time per-structure: Industry (+1.25% production), Fortify (+5 max life),
  Watchtower (+1 vision), Housing (+0.5 populace/turn), Treasury (+1 gold/turn).
- Player-wide techs: Military I (+1 HP to all Armies), Military II (+1 HP to all
  Tanks; requires Military I + 10 steel).
- Repeatable: Conscript (instant Army) and Repair (full heal).
- Production progress is now a double so the small Industry bonus accumulates.
- Costs tuned so the big player-wide techs are a real mid-game commitment.

Difficulty:
- Difficulty now scales AI resource income via GetAIBonusMultiplier
  (Easy 0.75x, Normal 1.0x = no advantage, Hard 1.25x, Expert 1.5x); the human is
  always 1.0x. Added the missing Expert option to the difficulty dropdown.

Start screen:
- Merged the splash branding (EMPIRE 2026 / J. Alessi / tagline) into the New Game
  screen header and removed the standalone splash window.

+++ Models/Base.cs
@@ -4,7 +4,7 @@
     public bool HasShipyard { get; set; }
     public Queue<UnitProductionOrder> ProductionQueue { get; set; }
     public int ProductionPointsPerTurn { get; set; }
-    public int CurrentProductionProgress { get; set; }
+    public double CurrentProductionProgress { get; set; }
     
     // Unit storage with capacity limits
     public List<AirUnit> Airport { get; set; }
